chore: Update npm dependencies and add @iconify/svelte and sass
This commit updates the npm dependencies in the package.json file to their latest stable versions. It also adds the @iconify/svelte package and the sass package as peer dependencies. These changes aim to improve the project by ensuring compatibility with the latest versions of the dependencies and adding support for using icons and Sass in the project.